requirements.txt

python-dotenv

spotipy

flask
#gunicorn # production

pytest
